correct user deletion

+++ app.py
@@ -314,21 +314,25 @@ def add_user():
     return render_template('add_user.html', form=form, name=name, our_users=our_users)
     
 @app.route('/delete/<int:id>')
+@login_required
 def delete(id):
-    user_to_delete = Users.query.get_or_404(id)
-    name = None
-    form = UserForm()
+    if id == current_user.id:
+        user_to_delete = Users.query.get_or_404(id)
+        name = None
+        form = UserForm()
 
-    try:
-        db.session.delete(user_to_delete)
-        db.session.commit()
-        flash("User Deleted Successfully")
-        our_users = Users.query.order_by(Users.date_added)
-        return render_template('add_user.html', form=form, name=name, our_users=our_users)
-    except:
-        flash("Something went wrong!") 
-        return render_template('add_user.html', form=form, name=name, our_users=our_users)
-
+        try:
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash("User Deleted Successfully")
+            our_users = Users.query.order_by(Users.date_added)
+            return render_template('add_user.html', form=form, name=name, our_users=our_users)
+        except:
+            flash("Something went wrong!") 
+            return render_template('add_user.html', form=form, name=name, our_users=our_users)
+    else:
+        flash("You can't delete another user!")
+        return redirect(url_for('dashboard'))
 #def index():
 #    return "<h1>Hello world</h1>"
